requirements.txt
boto3
pytesseract
pillow
